added security for checking that only an admin user can delete others…

+++ server/src/main/java/org/spring2885/server/api/PersonsApi.java
@@ -12,16 +12,19 @@ import org.spring2885.server.db.service.PersonService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.stereotype.Controller;
+import org.springframework.security.web.servletapi.SecurityContextHolderAwareRequestWrapper;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.RestController;
 
+import com.google.common.base.Preconditions;
 import com.google.common.collect.FluentIterable;
+import com.google.common.collect.Iterables;
 
-@Controller
+@RestController
 @RequestMapping(value = "/api/v1/profiles", produces = { APPLICATION_JSON_VALUE })
 public class PersonsApi {
 	
@@ -29,8 +32,8 @@ public class PersonsApi {
 	private PersonService personService;
 
 	@RequestMapping(value = "/{id}", method = RequestMethod.GET)
-	public ResponseEntity<Person> get(@PathVariable("id") int id)
-			throws NotFoundException {
+	public ResponseEntity<Person> get(
+			@PathVariable("id") int id) throws NotFoundException {
 		DbPerson o = personService.findById(id);
 		if (o == null) {
 			// When adding test testPersonsById_notFound, was getting a NullPointerException
@@ -41,8 +44,14 @@ public class PersonsApi {
 	}
 
 	@RequestMapping(value = "/{id}", method = RequestMethod.DELETE)
-	public ResponseEntity<Void> delete(@PathVariable("id") int id)
+	public ResponseEntity<Void> delete(
+			@PathVariable("id") Integer id,
+			SecurityContextHolderAwareRequestWrapper request
+			)
 			throws NotFoundException {
+		
+		checkAdminRequestIfNeeded(id, request);
+
 		personService.delete(id);
 		return new ResponseEntity<>(HttpStatus.OK);
 	}
@@ -60,8 +69,31 @@ public class PersonsApi {
 
 	// TODO(rcleveng): Implement this one.
 	@RequestMapping(method = RequestMethod.PUT)
-	public ResponseEntity<Void> put(@RequestBody Person person) throws NotFoundException {
+	public ResponseEntity<Void> put(
+			@RequestBody Person person,
+			SecurityContextHolderAwareRequestWrapper request) throws NotFoundException {
+		checkAdminRequestIfNeeded(person.getId().intValue(), request);
 		return new ResponseEntity<>(HttpStatus.OK);
+	}
+	
+	
+	private void checkAdminRequestIfNeeded(int id, SecurityContextHolderAwareRequestWrapper request) {
+		if (!request.isUserInRole("ROLE_ADMIN")) {
+			// Only admin's can change other profiles.
+			String name = request.getUserPrincipal().getName();
+			List<DbPerson> possibles = personService.findByEmail(name);
+			Preconditions.checkState(possibles.size() == 1, "I can't find myself... need more zen.");
+			
+			DbPerson me = Iterables.getOnlyElement(possibles);
+			if (me.getId() != id) {
+				// Someone is being naughty...
+				throw new RuntimeException("Only admin's can change others... Read this: "
+						+ "God, grant me the serenity to accept the things I cannot change,"
+						+ "Courage to change the things I can,"
+						+ "And wisdom to know the difference.");
+			}
+		}
+		
 	}
 
 }
